feat: add topics and participants to summary page
